Fix falcon_link step
We had a done file depend on a sub-task of fc_run, and a link
to that file depended on the same done file. That's a loop,
and it made restarts difficult.

We will need a cleaner solution someday, but this works for now.

However, some later step (maybe pbalign or chunking?) is not idempotent.
That's a separate issue.

SAT-46